moved router into router/index.ts
